Move default setSortProperties into sortUtils
